remove benchmark packages. include benchmarks from mandala into karura
