+++ Cargo.toml
@@ -13,7 +13,6 @@ members = [
 	"rpc",
 
 	"runtime/common",
-	"runtime/benchmarks",
 	"runtime/mandala",
 	"runtime/karura",
 	"runtime/acala",
