+++ runtime/mandala/Cargo.toml
@@ -122,7 +122,6 @@ module-staking-pool-rpc-runtime-api = { path = "../../modules/staking-pool/rpc/r
 module-polkadot-bridge = { path = "../../modules/polkadot-bridge", default-features = false }
 primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 runtime-common = { path = "../common", default-features = false }
-benchmarks = { path = "../benchmarks", default-features = false, optional = true }
 
 module-evm-rpc-runtime-api = { path = "../../modules/evm/rpc/runtime_api",  default-features = false }
 
@@ -269,7 +268,6 @@ std = [
 runtime-benchmarks = [
 	"frame-benchmarking",
 	"orml-benchmarking",
-	"benchmarks",
 	"libsecp256k1/hmac",
 
 	"frame-support/runtime-benchmarks",
